requirements.txt
pandas
pymupdf
#camelot-py
#ghostscript
tabula-py